fix(mobile): render PR compose/link forms inline in sidebar
The mobile create-PR and link-PR forms were built on BottomDrawer (a
full-screen absolute-fill overlay) but rendered inside the PR sidebar's
ScrollView. A nested absolute-fill overlay only fills the small empty-state
region, clipping the compose form right after the Base-branch field.

Extract the form bodies into plain-View components (MobilePrComposeForm,
MobileLinkPrForm) and render them inline in the sidebar empty state via a
local mode state. The BottomDrawer wrappers (MobilePrComposeSheet for the
full-screen source-control modals) now reuse the same form body instead of
duplicating it.
